requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
PyJWT==1.7.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
twilio==6.58.0
tzlocal==2.1
urllib3==1.26.4
